requirements.txt
asgiref==3.4.1
Django==3.2.13
django-filter==21.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
structlog==21.5.0
typing-extensions==4.1.1
